requirements.txt
flask
transformers
torch
pillow
jsonschema
scikit-learn
